Added files via upload
The before file shows the application as it is now. It has a lot of elements that might be frustrating and not intuitive for users including the placements of different buttons, overlapping content or the labels of different buttons. The Sketch shows a rough drawing of the initial idea we had after discussing the application in more detail and what everything should function like. And lastly, the prototype image is what the application will look like, it was created using Sketch.io.